fix(maintainability): Step 5 — move VS Code UI types out of model/
Move DecoratedQuickPickItem + TimedQuickPick from model/vscode.ts to
provider/features/scan-entries.ts; export via provider/index.ts.
Move TextMateRule inline to startup.ts (only consumer).
Delete src/model/vscode.ts; update J.Model refs in dialogues.ts to J.Provider.

#188
